# openapi.yaml
openapi: 3.0.3
info:
  title: ManadaBook API
  description: API completa para la red social de mascotas más grande del mundo
  version: 1.0.0
  contact:
    name: ManadaBook Support
    email: support@manadabook.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.manadabook.com
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Snippet:
      type: object
      properties:
        id:
          type: string
          format: uuid
        author_id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 500
        video_url:
          type: string
          format: uri
        thumbnail_url:
          type: string
          format: uri
        duration:
          type: integer
          minimum: 0
        views:
          type: integer
          minimum: 0
        likes:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
      required: [id, author_id, title, video_url]
...
paths:
...
  /snippets:
    get:
      tags:
        - Snippets
      summary: Listar videos cortos
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 10
      responses:
        '200':
          description: Lista de videos
          content:
            application/json:
              schema:
                type: object
                properties:
                  snippets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Snippet'
                  total:
                    type: integer
                  page:
                    type: integer
                  hasMore:
                    type: boolean

    post:
      tags:
        - Snippets
      summary: Subir nuevo video
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  maxLength: 200
                description:
                  type: string
                  maxLength: 500
                video:
                  type: string
                  format: binary
      responses:
        '201':
          description: Video subido exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Snippet'
...